requirements.txt
# Gemini API and Google Auth
google-generativeai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-genai
# Text-to-Speech and Video Editing
gTTS
moviepy==1.0.3

# For Pexels API
requests

# For image processing (e.g., slides, thumbnails)
Pillow==9.5.0

# ✅ For MP3 to WAV conversion
pydub